requirements.txt
jupyter==1.0.0
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
matplotlib==3.7.1
openai==0.27.8
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
pytest-benchmark==4.0.0
transformers==4.30.0
torch==2.0.1
langchain==0.0.198 
pydantic==1.10.9
python-dotenv==1.0.0
Flask==2.3.2
gradio==3.35.2